+++ src/commons/mod.rs
@@ -2,10 +2,12 @@
 
 pub mod affinity;
 pub mod authentication;
+pub mod cluster_operation;
 pub mod ldap;
 pub mod listener;
 pub mod opa;
 pub mod product_image_selection;
+pub mod rbac;
 pub mod resources;
 pub mod s3;
 pub mod secret_class;